Add files via upload
This code uses the GLC_FCS30 dataset in Google Earth Engine (GEE) to examine land cover change in a specific region of interest (ROI). It processes land cover images from 2000 to 2022 by reclassifying them into simplified categories, visualizing them on a map, and calculating the area of each land cover class for each year. The reclassification is accomplished by mapping the original class codes to new classes, and the class colours and labels are displayed in a legend. The script calculates the area (in square kilometres) of each land cover class annually and generates a time series for a specific class (for example, Class 34), displaying its change over time on a line chart.